move version parser to codyze-core
