+++ codyze-v3/codyze/src/main/kotlin/Main.kt
@@ -9,6 +9,7 @@ import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.*
 import de.fraunhofer.aisec.codyze.options.configFileOption
 import de.fraunhofer.aisec.codyze.source.JsonValueSource
+import de.fraunhofer.aisec.codyze_core.helper.VersionProvider
 import java.io.IOException
 import java.net.URL
 import java.nio.file.Path
@@ -47,59 +48,16 @@ class ConfigFileParser : CliktCommand(treatUnknownOptionsAsArgs = true) {
  */
 class CodyzeCli(val configFile: Path? = null) :
     CliktCommand(help = "Codyze finds security flaws in source code", printHelpOnEmptyArgs = true) {
+
+    /** Stores the version of Codyze */
+    private val versionProvider = VersionProvider("Codyze v3")
+
     init {
-        versionOption(Codyze.version, names = setOf("--version", "-V"))
+        versionOption(versionProvider.version, names = setOf("--version", "-V"))
         context {
             if (configFile != null)
                 valueSource = JsonValueSource.from(configFile, requireValid = true)
             helpFormatter = CliktHelpFormatter(showDefaultValues = true, requiredOptionMarker = "*")
-        }
-    }
-
-    /**
-     * Storage for constant application information.
-     *
-     * Note: Using `object` to keep information as singleton and prevent initialization on each
-     * object creation of [CodyzeCli].
-     */
-    private object Codyze {
-        /** Name of our application. */
-        private const val IMPLEMENTATION_TITLE: String = "Codyze v3"
-
-        /** Default version to use, when no explicit version has been set. */
-        private const val DEFAULT_VERSION: String = "0.0.0-SNAPSHOT"
-
-        /**
-         * Determine version from system property `codyze-v3-version` or from Codyzes' JAR file
-         * manifests.
-         *
-         * Note: Using `lazy` delegate to calculate property once, when it is needed. Reduces
-         * overhead by not reading in all manifests on the classpath multiple times.
-         */
-        val version: String by lazy {
-            System.getProperty("codyze-v3-version")
-                ?: run {
-                    val resources: Enumeration<URL> =
-                        CodyzeCli::class.java.classLoader.getResources("META-INF/MANIFEST.MF")
-                    while (resources.hasMoreElements()) {
-                        val url = resources.nextElement()
-                        try {
-                            val manifest = Manifest(url.openStream())
-                            val mainAttributes = manifest.mainAttributes
-                            if (
-                                IMPLEMENTATION_TITLE ==
-                                    mainAttributes.getValue(Attributes.Name.IMPLEMENTATION_TITLE)
-                            ) {
-                                return@run mainAttributes.getValue(
-                                    Attributes.Name.IMPLEMENTATION_VERSION
-                                )
-                            }
-                        } catch (ex: IOException) {
-                            logger.trace { "Unable to read from $url: $ex" }
-                        }
-                    }
-                    DEFAULT_VERSION
-                }
         }
     }
